fix: introduce codex-skills crate
