+++ codex-rs/skills/src/assets/samples/skill-creator/SKILL.md
