+++ codex-rs/skills/src/assets/samples/skill-creator/agents/openai.yaml
