+++ codex-rs/skills/src/assets/samples/skill-creator/assets/skill-creator.png
