+++ codex-rs/skills/src/assets/samples/skill-creator/license.txt
