+++ codex-rs/skills/src/assets/samples/skill-creator/references/openai_yaml.md
