+++ codex-rs/skills/src/assets/samples/skill-creator/scripts/generate_openai_yaml.py
