+++ codex-rs/skills/src/assets/samples/skill-creator/scripts/init_skill.py
