+++ codex-rs/skills/src/assets/samples/skill-creator/scripts/quick_validate.py
