+++ codex-rs/skills/src/assets/samples/skill-installer/LICENSE.txt
