+++ codex-rs/skills/src/assets/samples/skill-installer/SKILL.md
