+++ codex-rs/skills/src/assets/samples/skill-installer/agents/openai.yaml
