+++ codex-rs/skills/src/assets/samples/skill-installer/assets/skill-installer.png
