+++ codex-rs/skills/src/assets/samples/skill-installer/scripts/install-skill-from-github.py
