+++ codex-rs/skills/src/assets/samples/skill-installer/scripts/list-skills.py
